chore: Just check Dockerfile has correct tag after tagging
I realized that there is chicken and egg situation when the tag is
created. The workflow runs to update the Dockerfile but when the
auto-PR is merged it (obvious) didn't make it into the tag. For now
updating the tag in the Dockerfile needs to be done prior to tagging,
the CI will tell you if you haven't. At this point you'll need to
re-tag and correct the error.

Signed-off-by: crozzy <joseph.crosland@gmail.com>