update: add the ability to specify a http-timeout
Previously the timeout was set to 2 mins which is usually plenty for
most scenarios but some users want to be able to specify a custom
timeout for more latent environments. This adds a --http-timeout flag to
the `update` mode and sets it on the client used to request data from
security sources.

Signed-off-by: crozzy <joseph.crosland@gmail.com>